% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsHLAfiltR_functions.R
\name{check.system2_output}
\alias{check.system2_output}
\title{System Check}
\usage{
check.system2_output(system2_output, system2_error)
}
\arguments{
\item{system2_output}{The output of a command passed to the system2() function.}

\item{system2_error}{The message returned if the external resource is unavailable.}
}
\description{
Checks the exteral system for the avilability of external resources. If the resource is not available, execution of the package is halted.
}
\details{
For internal wgsHLAfiltR use.
}
\note{
This function is for interal use.
}
