% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsHLAfiltR_functions.R
\name{extractor.run}
\alias{extractor.run}
\title{Extract HLA Reads}
\usage{
extractor.run(sampleList, threads, extractedFastqDirectory, forceRun = F)
}
\arguments{
\item{sampleList}{The sampleList object returned by the general.paired_sample_objects() function.}

\item{threads}{The number of threads specified for bowtie2.}

\item{extractedFastqDirectory}{The directory into which filtered fastq files will be writen.}

\item{forceRun}{A logical value. If forceRun = TRUE, the alignment is forced to run, even if there are previous results.}
}
\value{
The populated sampleList object.
}
\description{
Extracts HLA reads based on bowtie2 alignment for each subject.
}
\details{
For internal wgsHLAfiltR use.
}
\note{
This function is for interal use.
}
