% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsHLAfiltR_functions.R
\name{filterHLA}
\alias{filterHLA}
\title{Main Function For Extracting Classical HLA Reads From fastq.gz Files}
\usage{
filterHLA(inputDirectory, outputDirectory = "Results")
}
\arguments{
\item{inputDirectory}{A path to the directory that contains the fastq.gz files to be filtered. This parameter is required.}

\item{outputDirectory}{A path to an existing directory where the HLA-specific fastq.gz files will be written. If no directory is specified, a 'Results' directory will be created in the working directory. If a 'Results' directory is already present in the working directory, output files will be written in that directory.}
}
\value{
A list object describing all of the pre-and post-filtering reads for each subject.
}
\description{
This function and calls all of the wgsHLAfiltR package's secondary functions, extracting HLA reads from fastq.gz files.
}
\details{
This function identifies whole-genome and whole-exome reads that map to the classical HLA loci (HLA-A, -C, -B, -DRB1, -DRB3/4/5, -DQA1, -DQB1, -DPA1, and -DPB1) in fastq.gz files in a user-specified directory, and writes fastq.gz files containing only classical HLA-specific reads in a second output directory. Reads for the HLA genes are mapped using hg19 reference coordinates.
}
\note{
This function requires that bowtie2 be installed in the local environment.
}
\examples{
# Filter non-HLA reads in fastq.gz files in a single directory
# inputDir <- paste(getwd(),"unfilteredReads",sep="/")
# outputDir <- paste(getwd(),"HLAfilteredReads",sep="/")
# readFilteringData <- filterHLA(inputDirectory=inputDir,outputDirectory=outputDir) 
}
\keyword{HLA}
\keyword{fastq}
\keyword{filter}
\keyword{reads}
