% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsHLAfiltR_functions.R
\name{general.paired_sample_objects}
\alias{general.paired_sample_objects}
\title{General Paired Sample Objects}
\usage{
general.paired_sample_objects(
  rawFastqDirectory,
  fastqPattern,
  resultsDirectory,
  shortNameDelim
)
}
\arguments{
\item{rawFastqDirectory}{The directory containing unfiltered fastq files.}

\item{fastqPattern}{A sub-string of the fastq file names that identifies the fastq files to be filtered. A value of 'fastq' specifies all fastq files.}

\item{resultsDirectory}{The directory into which the filtered fastq files should be written.}

\item{shortNameDelim}{A delimeter used to shorten fastq filenames.}
}
\value{
A list of sample objects that contain the paired fastq file names for each subject, identifying each fastq pair, the path to each paired file, and a logigical indicator of gzip status.
}
\description{
Finds pairs of fastq files inputDirectory for each subject, and returns a list of sample objects that contain the paired fastq file names
}
\details{
For internal wgsHLAfiltR use.
}
\note{
This function is for interal use.
}
