% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqSample}
\alias{seqSample}
\title{Sequenced Samples object class}
\format{
An object of class \code{refObjectGenerator} of length 1.
}
\usage{
seqSample
}
\description{
An object class function required for the package.
}
\keyword{datasets}
